+++ src/At.ts
@@ -41,6 +41,19 @@ export const fromIso = <T, S>(iso: Iso<T, S>) => <I, A>(sia: At<S, I, A>): At<T,
 
 /**
  * @category constructors
- * @since 2.3.2
+ * @since 2.3.7
  */
-export const atRecord: <A = never>() => At<Readonly<Record<string, A>>, string, Option<A>> = _.atRecord
+export const atReadonlyRecord: <A = never>() => At<Readonly<Record<string, A>>, string, Option<A>> = _.atReadonlyRecord
+
+// -------------------------------------------------------------------------------------
+// deprecated
+// -------------------------------------------------------------------------------------
+
+/**
+ * Use `atReadonlyRecord` instead.
+ *
+ * @category constructors
+ * @since 2.3.2
+ * @deprecated
+ */
+export const atRecord: <A = never>() => At<Readonly<Record<string, A>>, string, Option<A>> = _.atReadonlyRecord