+++ src/Optional.ts
@@ -216,7 +216,7 @@ export const key = (key: string) => <S, A>(sa: Optional<S, Readonly<Record<strin
  * @since 2.3.0
  */
 export const atKey = (key: string) => <S, A>(sa: Optional<S, Readonly<Record<string, A>>>): Optional<S, Option<A>> =>
-  pipe(sa, compose(_.lensAsOptional(_.atRecord<A>().at(key))))
+  pipe(sa, compose(_.lensAsOptional(_.atReadonlyRecord<A>().at(key))))
 
 /**
  * Return a `Optional` from a `Optional` focused on the `Some` of a `Option` type.